pipelines/datalake_autotagging/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
beartype==0.17.2
    # via picsellia
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecation==2.1.0
    # via picsellia
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fsspec==2024.9.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.25.1
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==2.1.1
    # via transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.68
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
orjson==3.10.7
    # via picsellia
packaging==24.1
    # via
    #   deprecation
    #   huggingface-hub
    #   transformers
picsellia==6.18.2
    # via -r requirements.in
picsellia-annotations==0.9.0
    # via picsellia
picsellia-connexion-services==0.2.1
    # via picsellia
pillow==10.4.0
    # via
    #   -r requirements.in
    #   picsellia
pydantic==2.9.2
    # via
    #   picsellia
    #   picsellia-annotations
pydantic-core==2.23.4
    # via pydantic
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   picsellia
    #   transformers
regex==2024.9.11
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   picsellia
    #   picsellia-connexion-services
    #   transformers
safetensors==0.4.5
    # via
    #   -r requirements.in
    #   transformers
sympy==1.13.3
    # via torch
tabulate==0.9.0
    # via -r requirements.in
tdqm==0.0.1
    # via picsellia
tokenizers==0.19.1
    # via transformers
torch==2.4.1
    # via -r requirements.in
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   tdqm
    #   transformers
transformers==4.44.2
    # via -r requirements.in
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.3
    # via requests
xmltodict==0.13.0
    # via picsellia-annotations
